Second update with some projects and other info
I decided to discuss more about the things I learned rather than all the projects since most projects were follow along and on the simpler side of things. Ultimately what I learned will be more applicable for job searching and the fact that I have a good foundation in C# to build on.